some modification on RegisterController

+++ WebshopRegApp/src/main/java/com/homework/webshopregapp/jpa/session/AbstractFacade.java
@@ -1,7 +1,14 @@
 package com.homework.webshopregapp.jpa.session;
 
+import com.homework.webshopregapp.jsf.util.JsfUtil;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
 import javax.persistence.EntityManager;
+import javax.validation.ConstraintViolation;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
 
 public abstract class AbstractFacade<T> {
     private Class<T> entityClass;
@@ -12,16 +19,41 @@ public abstract class AbstractFacade<T> {
 
     protected abstract EntityManager getEntityManager();
 
+    private boolean constraintValidationsDetected(T entity) {
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        Validator validator = factory.getValidator();
+        Set<ConstraintViolation<T>> constraintViolations = validator.validate(entity);
+        if (constraintViolations.size() > 0) {
+          Iterator<ConstraintViolation<T>> iterator = constraintViolations.iterator();
+          while (iterator.hasNext()) {
+            ConstraintViolation<T> cv = iterator.next();
+            System.err.println(cv.getRootBeanClass().getName() + "." + cv.getPropertyPath() + " " + cv.getMessage());
+
+            JsfUtil.addErrorMessage(cv.getRootBeanClass().getSimpleName() + "." + cv.getPropertyPath() + " " + cv.getMessage());
+          }
+          return true;
+        }
+        else {
+          return false;
+        }
+     }    
+        
     public void create(T entity) {
-        getEntityManager().persist(entity);
+        if (!constraintValidationsDetected(entity)) {        
+            getEntityManager().persist(entity);
+        }    
     }
 
     public void edit(T entity) {
-        getEntityManager().merge(entity);
+        if (!constraintValidationsDetected(entity)) {        
+            getEntityManager().merge(entity);
+        }         
     }
 
     public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+        if (!constraintValidationsDetected(entity)) {        
+            getEntityManager().remove(getEntityManager().merge(entity));
+        }    
     }
 
     public T find(Object id) {
